make layout and crud user, fixing image note api, fixing get note api

+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,7 @@ use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\Rule;
 
 class UserController extends Controller
 {
@@ -30,7 +31,7 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view("User.create");
+        return view("Users.create");
     }
 
     /**
@@ -41,6 +42,29 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        // validating data
+        $validatedData = $request->validate([
+            'name' => 'required|max:255',
+            'email' => 'required|unique:users,email|max:255',
+            'phone_number' => 'required|numeric',
+            'password' => 'required',
+            'confirm_password' => 'required|same:password',
+            'image' => 'nullable|image'
+        ]);
+
+        // storing image
+        if ($request->image !== null) {
+            $validatedData['image'] = $request->file('image')->store('images');
+        } else {
+            $validatedData['image'] = 'images/default.png';
+        }
+
+        // hashing password
+        $validatedData['password'] = Hash::make($request->password);
+
+        // create user
+        User::create($validatedData);
+        return redirect('mynotes-users')->with('success', 'Data Created Successfully');
     }
 
     /**
@@ -52,7 +76,7 @@ class UserController extends Controller
     public function show($id)
     {
         $user = User::findOrFail($id);
-        return view("User.show", [
+        return view("Users.show", [
             'user' => $user
         ]);
     }
@@ -66,7 +90,7 @@ class UserController extends Controller
     public function edit($id)
     {
         $user = User::findOrFail($id);
-        return view("User.edit", [
+        return view("Users.edit", [
             'user' => $user
         ]);
     }
@@ -80,6 +104,35 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // validating password
+        $validatedData = $request->validate([
+            'name' => 'required|max:255',
+            'email' => [Rule::unique('users')->ignore($id), 'nullable', 'max:255', 'email'],
+            'phone_number' => 'required|numeric',
+            'password' => 'required',
+            'confirm_password' => 'required|same:password',
+            'image' => 'nullable|image'
+        ]);
+
+        // find user
+        $user = User::findOrFail($id);
+
+        // storing image
+        if ($request->image !== null) {
+            Storage::delete($user->image);
+            $validatedData['image'] = $request->file('image')->store('images');
+        } else {
+            $validatedData['image'] = $user->image;
+        }
+
+        // hashing password
+        $validatedData['password'] = Hash::make($request->password);
+
+        // update user
+        $user->update($validatedData);
+
+        // return view
+        return redirect('mynotes-users')->with('success', 'Data Updated Successfully');
     }
 
     /**
@@ -91,7 +144,9 @@ class UserController extends Controller
     public function destroy($id)
     {
         $user = User::findOrFail($id);
-        Storage::delete($user->image);
+        if ($user->image !== 'images/default.png') {
+            Storage::delete($user->image);
+        }
         $user->delete();
         return back()->with('success', 'Data Deleted Successfully');
     }
